+++ tools/corrosion/cmake/Corrosion.cmake
@@ -53,6 +53,54 @@ get_property(
     TARGET Rust::Cargo PROPERTY IMPORTED_LOCATION
 )
 
+if(Rust_TOOLCHAIN_IS_RUSTUP_MANAGED AND DEFINED Rust_RUSTUP_TOOLCHAINS)
+    set(corrosion_tools_rust_toolchain_docstring "Rust toolchain to use for building helper tools such as cbindgen or cxx-bridge")
+    if(DEFINED CORROSION_TOOLS_RUST_TOOLCHAIN)
+        set(cor_default_tools_toolchain "${CORROSION_TOOLS_RUST_TOOLCHAIN}")
+    else()
+        set(cor_default_tools_toolchain "${Rust_TOOLCHAIN}")
+    endif()
+    set(CORROSION_TOOLS_RUST_TOOLCHAIN "${cor_default_tools_toolchain}" CACHE STRING
+        "${corrosion_tools_rust_toolchain_docstring}" FORCE)
+    set_property(CACHE CORROSION_TOOLS_RUST_TOOLCHAIN PROPERTY STRINGS "${Rust_RUSTUP_TOOLCHAINS}")
+    if(NOT "$CACHE{CORROSION_TOOLS_RUST_TOOLCHAIN}" IN_LIST Rust_RUSTUP_TOOLCHAINS)
+        if("$CACHE{CORROSION_TOOLS_RUST_TOOLCHAIN}-${Rust_CARGO_HOST_TARGET}" IN_LIST Rust_RUSTUP_TOOLCHAINS)
+            set(CORROSION_TOOLS_RUST_TOOLCHAIN "$CACHE{CORROSION_TOOLS_RUST_TOOLCHAIN}-${Rust_CARGO_HOST_TARGET}"
+                CACHE PATH "${corrosion_tools_rust_toolchain_docstring}" FORCE)
+        else()
+            message(FATAL_ERROR "CORROSION_TOOLS_RUST_TOOLCHAIN must be set to a valid rustup managed toolchain path."
+                    "Rust_RUSTUP_TOOLCHAINS contains a list of valid installed toolchains."
+            )
+        endif()
+    endif()
+    foreach(toolchain tc_rustc tc_cargo IN ZIP_LISTS Rust_RUSTUP_TOOLCHAINS Rust_RUSTUP_TOOLCHAINS_RUSTC_PATH Rust_RUSTUP_TOOLCHAINS_CARGO_PATH)
+        if("${toolchain}" STREQUAL $CACHE{CORROSION_TOOLS_RUST_TOOLCHAIN})
+            # Minimum CMake version 3.29 for `IS_EXECUTABLE`.
+            if(NOT (tc_cargo AND tc_rustc ))
+                message(FATAL_ERROR "Failed to find executable rustc or cargo for toolchain `$CACHE{CORROSION_TOOLS_RUST_TOOLCHAIN}`")
+            endif()
+            set(CORROSION_TOOLS_RUSTC "${tc_rustc}" CACHE INTERNAL "" FORCE)
+            set(CORROSION_TOOLS_CARGO "${tc_cargo}" CACHE INTERNAL "" FORCE)
+            break()
+        endif()
+    endforeach()
+    if(NOT DEFINED CACHE{CORROSION_TOOLS_CARGO})
+        message(FATAL_ERROR "Internal error: Failed to find toolchain $CACHE{CORROSION_TOOLS_RUST_TOOLCHAIN} in "
+                "list of rustup managed toolchains: ${Rust_RUSTUP_TOOLCHAINS}"
+        )
+    endif()
+else()
+    # Fallback to the default project toolchain if rust is not rustup managed.
+    if(DEFINED CORROSION_TOOLS_RUST_TOOLCHAIN)
+        message(DEBUG "Ignoring `CORROSION_TOOLS_RUST_TOOLCHAIN=${CORROSION_TOOLS_RUST_TOOLCHAIN}` "
+            "since the toolchains are not rustup managed. Falling back to the default rust toolchain"
+            " for this project."
+        )
+    endif()
+    set(CORROSION_TOOLS_RUSTC "${RUSTC_EXECUTABLE}" CACHE INTERNAL "" FORCE)
+    set(CORROSION_TOOLS_CARGO "${CARGO_EXECUTABLE}" CACHE INTERNAL "" FORCE)
+endif()
+
 function(_corrosion_bin_target_suffix target_name out_var_suffix)
     get_target_property(hostbuild "${target_name}" ${_CORR_PROP_HOST_BUILD})
     if((hostbuild AND CMAKE_HOST_WIN32)
@@ -64,6 +112,26 @@ function(_corrosion_bin_target_suffix target_name out_var_suffix)
         set(_suffix "")
     endif()
     set(${out_var_suffix} "${_suffix}" PARENT_SCOPE)
+endfunction()
+
+function(_handle_output_directory_genex input_path config_type output_path)
+    if("${config_type}" STREQUAL "")
+        # Prevent new path from being `dir//file`, since that causes issues with the
+        # file dependency.
+        string(REPLACE "/\$<CONFIG>" "${config_type}" curr_out_dir "${input_path}")
+        string(REPLACE "\$<CONFIG>" "${config_type}" curr_out_dir "${curr_out_dir}")
+    else()
+        string(REPLACE "\$<CONFIG>" "${config_type}" curr_out_dir "${input_path}")
+    endif()
+    string(GENEX_STRIP "${curr_out_dir}" stripped_out_dir)
+    if("${stripped_out_dir}" STREQUAL "${curr_out_dir}")
+        set("${output_path}" "${curr_out_dir}" PARENT_SCOPE)
+    else()
+        unset("${output_path}" PARENT_SCOPE)
+        message(WARNING "Encountered output directory path with unsupported genex. "
+                "Output dir: `${curr_out_dir}`"
+                "Note: Corrosion only supports the `\$<CONFIG>` generator expression for output directories.")
+    endif()
 endfunction()
 
 # Do not call this function directly!
@@ -100,16 +168,19 @@ function(_corrosion_set_imported_location_deferred target_name base_property out
         if(output_dir_curr_config)
             set(curr_out_dir "${output_dir_curr_config}")
         elseif(output_directory)
-            set(curr_out_dir "${output_directory}")
+            string(GENEX_STRIP "${output_directory}" output_dir_no_genex)
+            # Only add config dir if there is no genex in here. See
+            # https://cmake.org/cmake/help/latest/prop_tgt/RUNTIME_OUTPUT_DIRECTORY.html
+            if(output_directory STREQUAL output_dir_no_genex)
+                set(curr_out_dir "${output_directory}/${config_type}")
+            else()
+                set(curr_out_dir "${output_directory}")
+            endif()
         else()
             set(curr_out_dir "${CMAKE_CURRENT_BINARY_DIR}")
         endif()
-        string(REPLACE "\$<CONFIG>" "${config_type}" curr_out_dir "${curr_out_dir}")
-        message(DEBUG "Setting ${base_property}_${config_type_upper} for target ${target_name}"
-                " to `${curr_out_dir}/${filename}`.")
-
-        string(GENEX_STRIP "${curr_out_dir}" stripped_out_dir)
-        if(NOT ("${stripped_out_dir}" STREQUAL "${curr_out_dir}"))
+        _handle_output_directory_genex("${curr_out_dir}" "${config_type}" sanitized_out_dir)
+        if(NOT DEFINED sanitized_out_dir)
             message(FATAL_ERROR "${output_directory_property} for target ${output_dir_prop_target_name} "
                     "contained an unexpected Generator expression. Output dir: `${curr_out_dir}`"
                 "Note: Corrosion only supports the `\$<CONFIG>` generator expression for output directories.")
@@ -119,9 +190,9 @@ function(_corrosion_set_imported_location_deferred target_name base_property out
         set_property(
             TARGET ${target_name}
             PROPERTY "${base_property}_${config_type_upper}"
-                "${curr_out_dir}/${filename}"
+                "${sanitized_out_dir}/${filename}"
         )
-        set(base_output_directory "${curr_out_dir}")
+        set(base_output_directory "${sanitized_out_dir}")
     endforeach()
 
     if(NOT COR_IS_MULTI_CONFIG)
@@ -130,13 +201,13 @@ function(_corrosion_set_imported_location_deferred target_name base_property out
         else()
             set(base_output_directory "${CMAKE_CURRENT_BINARY_DIR}")
         endif()
-        string(REPLACE "\$<CONFIG>" "${CMAKE_BUILD_TYPE}" base_output_directory "${base_output_directory}")
-        string(GENEX_STRIP "${base_output_directory}" stripped_out_dir)
-        if(NOT ("${stripped_out_dir}" STREQUAL "${base_output_directory}"))
+        _handle_output_directory_genex("${base_output_directory}" "${CMAKE_BUILD_TYPE}" sanitized_output_directory)
+        if(NOT DEFINED sanitized_output_directory)
             message(FATAL_ERROR "${output_dir_prop_target_name} for target ${output_dir_prop_target_name} "
-                    "contained an unexpected Generator expression. Output dir: `${base_output_directory}`"
+                    "contained an unexpected Generator expression. Output dir: `${base_output_directory}`."
                     "Note: Corrosion only supports the `\$<CONFIG>` generator expression for output directories.")
         endif()
+        set(base_output_directory "${sanitized_output_directory}")
     endif()
 
     message(DEBUG "Setting ${base_property} for target ${target_name}"
@@ -204,10 +275,15 @@ function(_corrosion_copy_byproduct_deferred target_name output_dir_prop_names ca
         if(output_dir_curr_config)
             set(curr_out_dir "${output_dir_curr_config}")
         elseif(output_dir)
-            # Fallback to `output_dir` if specified
-            # Note: Multi-configuration generators append a per-configuration subdirectory to the
-            # specified directory unless a generator expression is used (from CMake documentation).
-            set(curr_out_dir "${output_dir}")
+            string(GENEX_STRIP "${output_dir}" output_dir_no_genex)
+            # Only add config dir if there is no genex in here. See
+            # https://cmake.org/cmake/help/latest/prop_tgt/RUNTIME_OUTPUT_DIRECTORY.html
+            # Logic duplicated from _corrosion_set_imported_location_deferred
+            if(output_dir STREQUAL output_dir_no_genex)
+                set(curr_out_dir "${output_dir}/${config_type}")
+            else()
+                set(curr_out_dir "${output_dir}")
+            endif()
         else()
             # Fallback to the default directory. We do not append the configuration directory here
             # and instead let CMake do this, since otherwise the resolving of dynamic library
@@ -582,11 +658,11 @@ function(_add_cargo_build out_cargo_build_out_dir)
     get_filename_component(workspace_toml_dir ${path_to_toml} DIRECTORY )
 
     if (CMAKE_VS_PLATFORM_NAME)
-        set (build_dir "${CMAKE_VS_PLATFORM_NAME}/$<CONFIG>")
+        set(build_dir "${CMAKE_VS_PLATFORM_NAME}/$<CONFIG>")
     elseif(COR_IS_MULTI_CONFIG)
-        set (build_dir "$<CONFIG>")
+        set(build_dir "$<CONFIG>")
     else()
-        set (build_dir .)
+        unset(build_dir)
     endif()
 
     # If a CMake sysroot is specified, forward it to the linker rustc invokes, too. CMAKE_SYSROOT is documented
@@ -600,6 +676,11 @@ function(_add_cargo_build out_cargo_build_out_dir)
     endif()
     if(COR_NO_DEFAULT_FEATURES)
         set(no_default_features_arg --no-default-features)
+    endif()
+    if(COR_NO_USES_TERMINAL)
+        unset(cor_uses_terminal)
+    else()
+        set(cor_uses_terminal USES_TERMINAL)
     endif()
 
     set(global_rustflags_target_property "$<TARGET_GENEX_EVAL:${target_name},$<TARGET_PROPERTY:${target_name},INTERFACE_CORROSION_RUSTFLAGS>>")
@@ -626,7 +707,8 @@ function(_add_cargo_build out_cargo_build_out_dir)
     set(cargo_target_option "--target=$<IF:${hostbuild_override},${_CORROSION_RUST_CARGO_HOST_TARGET},${_CORROSION_RUST_CARGO_TARGET}>")
 
     # The target may be a filepath to custom target json file. For host targets we assume that they are built-in targets.
-    _corrosion_strip_target_triple(${_CORROSION_RUST_CARGO_TARGET} stripped_target_triple)
+    _corrosion_strip_target_triple("${_CORROSION_RUST_CARGO_TARGET}" stripped_target_triple)
+    _corrosion_strip_target_triple("${_CORROSION_RUST_CARGO_TARGET_UPPER}" stripped_target_triple_upper)
     set(target_artifact_dir "$<IF:${hostbuild_override},${_CORROSION_RUST_CARGO_HOST_TARGET},${stripped_target_triple}>")
 
     set(flags_genex "$<GENEX_EVAL:$<TARGET_PROPERTY:${target_name},INTERFACE_CORROSION_CARGO_FLAGS>>")
@@ -674,7 +756,17 @@ function(_add_cargo_build out_cargo_build_out_dir)
     set(default_build_type_dir "$<IF:$<OR:$<CONFIG:Debug>,$<CONFIG:>>,debug,release>")
     set(build_type_dir "$<IF:${cargo_profile_set},${custom_profile_build_type_dir},${default_build_type_dir}>")
 
-    set(cargo_target_dir "${CMAKE_BINARY_DIR}/${build_dir}/cargo/build")
+    # We set a target folder based on the manifest path so if you build multiple workspaces (or standalone projects
+    # without workspace) they won't collide if they use a common dependency. This would confuse cargo and trigger
+    # unnecessary rebuilds
+    cmake_path(GET workspace_manifest_path PARENT_PATH parent_path)
+    cmake_path(GET parent_path PARENT_PATH grandparent_path)
+    string(REPLACE "${grandparent_path}/" "" cargo_folder_name "${parent_path}")
+    string(SHA1 cargo_path_hash ${workspace_manifest_path})
+    # Include a hash of the full path in case there are multiple projects with the same folder name
+    string(SUBSTRING "${cargo_path_hash}" 0 5 cargo_path_hash)
+    cmake_path(APPEND CMAKE_BINARY_DIR ${build_dir} cargo "${cargo_folder_name}_${cargo_path_hash}"
+               OUTPUT_VARIABLE cargo_target_dir)
     set(cargo_build_dir "${cargo_target_dir}/${target_artifact_dir}/${build_type_dir}")
     set("${out_cargo_build_out_dir}" "${cargo_build_dir}" PARENT_SCOPE)
 
@@ -684,15 +776,15 @@ function(_add_cargo_build out_cargo_build_out_dir)
         # This variable is read by cc-rs (often used in build scripts) to determine the c-compiler.
         # It can still be overridden if the user sets the non underscore variant via the environment variables
         # on the target.
-        list(APPEND corrosion_cc_rs_flags "CC_${_CORROSION_RUST_CARGO_TARGET_UNDERSCORE}=${CMAKE_C_COMPILER}")
+        list(APPEND corrosion_cc_rs_flags "CC_${stripped_target_triple}=${CMAKE_C_COMPILER}")
     endif()
     if(CMAKE_CXX_COMPILER)
-        list(APPEND corrosion_cc_rs_flags "CXX_${_CORROSION_RUST_CARGO_TARGET_UNDERSCORE}=${CMAKE_CXX_COMPILER}")
+        list(APPEND corrosion_cc_rs_flags "CXX_${stripped_target_triple}=${CMAKE_CXX_COMPILER}")
     endif()
     # cc-rs doesn't seem to support `llvm-ar` (commandline syntax), wo we might as well just use
     # the default AR.
     if(CMAKE_AR AND NOT (Rust_CARGO_TARGET_ENV STREQUAL "msvc"))
-        list(APPEND corrosion_cc_rs_flags "AR_${_CORROSION_RUST_CARGO_TARGET_UNDERSCORE}=${CMAKE_AR}")
+        list(APPEND corrosion_cc_rs_flags "AR_${stripped_target_triple}=${CMAKE_AR}")
     endif()
 
     # Since we instruct cc-rs to use the compiler found by CMake, it is likely one that requires also
@@ -731,7 +823,7 @@ function(_add_cargo_build out_cargo_build_out_dir)
         set(default_linker "$<IF:$<BOOL:${target_uses_cxx}>,${CMAKE_CXX_COMPILER},${CMAKE_C_COMPILER}>")
     endif()
     # Used to set a linker for a specific target-triple.
-    set(cargo_target_linker_var "CARGO_TARGET_${_CORROSION_RUST_CARGO_TARGET_UPPER}_LINKER")
+    set(cargo_target_linker_var "CARGO_TARGET_${stripped_target_triple_upper}_LINKER")
     set(linker "$<IF:${explicit_linker_defined},${explicit_linker_property},${default_linker}>")
     set(cargo_target_linker $<$<BOOL:${linker}>:${cargo_target_linker_var}=${linker}>)
 
@@ -746,6 +838,7 @@ function(_add_cargo_build out_cargo_build_out_dir)
     endif()
 
     message(DEBUG "TARGET ${target_name} produces byproducts ${build_byproducts}")
+    message(DEBUG "corrosion_cc_rs_flags: ${corrosion_cc_rs_flags}")
 
     add_custom_target(
         _cargo-build_${target_name}
@@ -785,7 +878,7 @@ function(_add_cargo_build out_cargo_build_out_dir)
         # The build is conducted in the directory of the Manifest, so that configuration files such as
         # `.cargo/config.toml` or `toolchain.toml` are applied as expected.
         WORKING_DIRECTORY "${workspace_toml_dir}"
-        USES_TERMINAL
+        ${cor_uses_terminal}
         COMMAND_EXPAND_LISTS
         VERBATIM
     )
@@ -813,9 +906,9 @@ function(_add_cargo_build out_cargo_build_out_dir)
         cargo-clean_${target_name}
         COMMAND
             "${cargo_bin}" clean ${cargo_target_option}
-            -p ${package_name} --manifest-path ${path_to_toml}
-        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/${build_dir}
-        USES_TERMINAL
+            -p ${package_name} --manifest-path "${path_to_toml}"
+        WORKING_DIRECTORY "${CMAKE_BINARY_DIR}/${build_dir}"
+        ${cor_uses_terminal}
     )
 
     if (NOT TARGET cargo-clean)
@@ -833,6 +926,7 @@ corrosion_import_crate(
         [NO_DEFAULT_FEATURES]
         [NO_STD]
         [NO_LINKER_OVERRIDE]
+        [NO_USES_TERMINAL]
         [LOCKED]
         [FROZEN]
         [PROFILE <cargo-profile>]
@@ -849,6 +943,7 @@ corrosion_import_crate(
 * **NO_DEFAULT_FEATURES**: Equivalent to [--no-default-features] passed to cargo build
 * **NO_STD**:  Disable linking of standard libraries (required for no_std crates).
 * **NO_LINKER_OVERRIDE**: Will let Rust/Cargo determine which linker to use instead of corrosion (when linking is invoked by Rust)
+* **NO_USES_TERMINAL**: Don't pass the `USES_TERMINAL` flag when creating the custom CMake targets.
 * **LOCKED**: Pass [`--locked`] to cargo build and cargo metadata.
 * **FROZEN**: Pass [`--frozen`] to cargo build and cargo metadata.
 * **PROFILE**: Specify cargo build profile (`dev`/`release` or a [custom profile]; `bench` and `test` are not supported)
@@ -873,7 +968,14 @@ corrosion_import_crate(
 ANCHOR_END: corrosion-import-crate
 #]=======================================================================]
 function(corrosion_import_crate)
-    set(OPTIONS ALL_FEATURES NO_DEFAULT_FEATURES NO_STD NO_LINKER_OVERRIDE LOCKED FROZEN)
+    set(OPTIONS
+        ALL_FEATURES
+        NO_DEFAULT_FEATURES
+        NO_STD
+        NO_LINKER_OVERRIDE
+        NO_USES_TERMINAL
+        LOCKED
+        FROZEN)
     set(ONE_VALUE_KEYWORDS MANIFEST_PATH PROFILE IMPORTED_CRATES)
     set(MULTI_VALUE_KEYWORDS CRATE_TYPES CRATES FEATURES FLAGS OVERRIDE_CRATE_TYPE)
     cmake_parse_arguments(COR "${OPTIONS}" "${ONE_VALUE_KEYWORDS}" "${MULTI_VALUE_KEYWORDS}" ${ARGN})
@@ -1247,6 +1349,7 @@ function(corrosion_install)
 
             # Gather the arguments to this install type
             set(ARGS)
+            list(LENGTH ARGN ARGN_LENGTH)
             while(ARGN_LENGTH)
                 # If the next keyword is an install target type, then break - arguments have been
                 # gathered.
@@ -1514,6 +1617,56 @@ set_target_properties(${INSTALL_TARGET}-shared
     endif()
 endfunction()
 
+function(_corrosion_check_cxx_version_helper manifest_dir cxx_name out_required_version)
+    execute_process(COMMAND ${CMAKE_COMMAND} -E env
+                    "CARGO_BUILD_RUSTC=${_CORROSION_RUSTC}"
+                    ${_CORROSION_CARGO} tree -i "${cxx_name}"
+                    # Usage of `cxx` could be gated behind a feature. Features can use Generator expressions,
+                    # so we can't really know what features we will enable when building at this point.
+                    # Features should be additive though, so simply enabling all-features should work for
+                    # dependency resolution.
+                    --all-features
+                    --target all
+                    --depth=0
+                    WORKING_DIRECTORY "${manifest_dir}"
+                    RESULT_VARIABLE cxx_version_result
+                    OUTPUT_VARIABLE cxx_version_output
+                    ERROR_VARIABLE cxx_version_error
+    )
+    if(NOT "${cxx_version_result}" EQUAL "0")
+        message(DEBUG "`cargo tree -i ${cxx_name}` returned an error: ${cxx_version_error}")
+        set("${out_required_version}" "${cxx_name}-NOTFOUND" PARENT_SCOPE)
+        return()
+    endif()
+    if(cxx_version_output MATCHES "${cxx_name} v([0-9]+.[0-9]+.[0-9]+)")
+        set("${out_required_version}" "${CMAKE_MATCH_1}" PARENT_SCOPE)
+    else()
+        message(DEBUG "Failed to parse `cargo tree -i ${cxx_name}` output: ${cxx_version_output}")
+        set("${out_required_version}" "${cxx_name}-NOTFOUND" PARENT_SCOPE)
+    endif()
+endfunction()
+
+function(_corrosion_check_cxx_version manifest_dir out_required_version)
+    # cxxbridge-cmd is known to be available in lockfiles since cxx 1.0.131.
+    # We include `cxx` as a fallback to support older versions too. `cxxbridge` should always
+    # be exactly the same version as `cxx`, so falling back to `cxx` version should not cause issues.
+    foreach(cxxbridge_name cxxbridge-cmd cxx)
+        unset(cxx_required_version)
+        _corrosion_check_cxx_version_helper("${manifest_dir}"
+                                            "${cxxbridge_name}"
+                                            cxx_required_version)
+        if(cxx_required_version)
+            set("${out_required_version}" "${cxx_required_version}" PARENT_SCOPE)
+            break()
+        else()
+            set("${out_required_version}" "cxx-NOTFOUND" PARENT_SCOPE)
+        endif()
+    endforeach()
+
+endfunction()
+
+
+
 #[=======================================================================[.md:
 ** EXPERIMENTAL **: This function is currently still considered experimental
   and is not officially released yet. Feedback and Suggestions are welcome.
@@ -1585,6 +1738,13 @@ function(corrosion_add_cxxbridge cxx_target)
         endif()
     endforeach()
 
+    if(DEFINED _arg_UNPARSED_ARGUMENTS)
+        message(AUTHOR_WARNING "corrosion_add_cxxbridge was called with the following unknown arguments: "
+                "`${_arg_UNPARSED_ARGUMENTS}`\n"
+                "Unknown arguments will be ignored."
+        )
+    endif()
+
     get_target_property(manifest_path "${_arg_CRATE}" INTERFACE_COR_PACKAGE_MANIFEST_PATH)
 
     if(NOT EXISTS "${manifest_path}")
@@ -1593,20 +1753,12 @@ function(corrosion_add_cxxbridge cxx_target)
 
     get_filename_component(manifest_dir ${manifest_path} DIRECTORY)
 
-    execute_process(COMMAND ${CMAKE_COMMAND} -E env
-        "CARGO_BUILD_RUSTC=${_CORROSION_RUSTC}"
-        ${_CORROSION_CARGO} tree -i cxx --depth=0
-        WORKING_DIRECTORY "${manifest_dir}"
-        RESULT_VARIABLE cxx_version_result
-        OUTPUT_VARIABLE cxx_version_output
-    )
-    if(NOT "${cxx_version_result}" EQUAL "0")
-        message(FATAL_ERROR "Crate ${_arg_CRATE} does not depend on cxx.")
-    endif()
-    if(cxx_version_output MATCHES "cxx v([0-9]+.[0-9]+.[0-9]+)")
-        set(cxx_required_version "${CMAKE_MATCH_1}")
-    else()
-        message(FATAL_ERROR "Failed to parse cxx version from cargo tree output: `cxx_version_output`")
+    _corrosion_check_cxx_version("${manifest_dir}" cxx_required_version)
+
+    if(NOT cxx_required_version)
+        message(FATAL_ERROR
+                "Failed to find a dependency on `cxxbridge-cmd` / `cxx` for crate ${_arg_CRATE}"
+        )
     endif()
 
     # First check if a suitable version of cxxbridge is installed
@@ -1640,19 +1792,21 @@ function(corrosion_add_cxxbridge cxx_target)
             if(Rust_CARGO_HOST_OS STREQUAL "windows")
                 set(executable_postfix ".exe")
             endif()
+
             add_custom_command(OUTPUT "${CMAKE_BINARY_DIR}/corrosion/cxxbridge_v${cxx_required_version}/bin/cxxbridge${executable_postfix}"
                 COMMAND
                 ${CMAKE_COMMAND} -E make_directory "${CMAKE_BINARY_DIR}/corrosion/cxxbridge_v${cxx_required_version}"
                 COMMAND
                     ${CMAKE_COMMAND} -E env
-                        "CARGO_BUILD_RUSTC=${_CORROSION_RUSTC}"
-                    ${_CORROSION_CARGO} install
+                        "CARGO_BUILD_RUSTC=$CACHE{CORROSION_TOOLS_RUSTC}"
+                    $CACHE{CORROSION_TOOLS_CARGO} install
                     cxxbridge-cmd
                     --version "${cxx_required_version}"
+                    --locked
                     --root "${CMAKE_BINARY_DIR}/corrosion/cxxbridge_v${cxx_required_version}"
                     --quiet
                     # todo: use --target-dir to potentially reuse artifacts
-                COMMENT "Building cxxbridge (version ${cxx_required_version})"
+                COMMENT "Building cxxbridge (version ${cxx_required_version}) with Rust toolchain $CACHE{CORROSION_TOOLS_RUST_TOOLCHAIN}"
                 )
             add_custom_target("cxxbridge_v${cxx_required_version}"
                 DEPENDS "${CMAKE_BINARY_DIR}/corrosion/cxxbridge_v${cxx_required_version}/bin/cxxbridge${executable_postfix}"
@@ -1695,14 +1849,11 @@ function(corrosion_add_cxxbridge cxx_target)
     # cxx generated code is using c++11 features in headers, so propagate c++11 as minimal requirement
     target_compile_features(${cxx_target} PUBLIC cxx_std_11)
 
-    # Todo: target_link_libraries is only necessary for rust2c projects.
-    # It is possible that checking if the rust crate is an executable is a sufficient check,
-    # but some more thought may be needed here.
-    # Maybe we should also let the user do this, since for c2rust, the user also has to call
-    # corrosion_link_libraries() themselves.
-    get_target_property(crate_target_type ${_arg_CRATE} TYPE)
-    if (NOT crate_target_type STREQUAL "EXECUTABLE")
-        target_link_libraries(${cxx_target} PRIVATE ${_arg_CRATE})
+    if (TARGET "${_arg_CRATE}-static")
+        target_link_libraries(${cxx_target} PRIVATE "${_arg_CRATE}-static")
+    endif()
+    if (TARGET "${_arg_CRATE}-shared")
+        target_link_libraries(${cxx_target} PRIVATE "${_arg_CRATE}-shared")
     endif()
 
     file(MAKE_DIRECTORY "${generated_dir}/include/rust")
@@ -1714,7 +1865,8 @@ function(corrosion_add_cxxbridge cxx_target)
             COMMENT "Generating rust/cxx.h header"
     )
 
-    set(GENERATED_FILES "${generated_dir}/include/rust/cxx.h")
+    set(GENERATED_SOURCES "")
+    set(GENERATED_HEADERS "${generated_dir}/include/rust/cxx.h")
 
     foreach(filepath ${_arg_FILES})
         get_filename_component(filename ${filepath} NAME_WE)
@@ -1746,15 +1898,21 @@ function(corrosion_add_cxxbridge cxx_target)
             COMMENT "Generating cxx bindings for crate ${_arg_CRATE} and file src/${filepath}"
         )
 
-        list(APPEND GENERATED_FILES
-            "${header_placement_dir}/${cxx_header}"
-            "${source_placement_dir}/${cxx_source}")
+        list(APPEND GENERATED_SOURCES "${source_placement_dir}/${cxx_source}")
+        list(APPEND GENERATED_HEADERS "${header_placement_dir}/${cxx_header}")
     endforeach()
-    target_sources(${cxx_target} PRIVATE ${GENERATED_FILES})
+    target_sources(${cxx_target} PRIVATE ${GENERATED_SOURCES})
+    # Make sure to export the headers with PUBLIC.
+    # This ensures that any target that depends on cxx_target also has these files as a dependency
+    # CMake will then make sure to generate the files before building either target, which is important
+    # in the presence of circular dependencies
+    target_sources(${cxx_target} PUBLIC ${GENERATED_HEADERS})
 
     if(DEFINED _arg_REGEN_TARGET)
+        # Add only the headers to the regen target, as the sources are actually not needed
+        # For the IDE to pick everything up
         add_custom_target(${_arg_REGEN_TARGET}
-            DEPENDS ${GENERATED_FILES}
+            DEPENDS ${GENERATED_HEADERS}
             COMMENT "Generated cxx bindings for crate ${_arg_CRATE}")
     endif()
 
@@ -1762,37 +1920,77 @@ endfunction()
 
 #[=======================================================================[.md:
 ANCHOR: corrosion_cbindgen
-```cmake
-corrosion_cbindgen(
-        TARGET <imported_target_name>
-        HEADER_NAME <output_header_name>
-        [CARGO_PACKAGE <cargo_package_name>]
-        [MANIFEST_DIRECTORY <package_manifest_directory>]
-        [CBINDGEN_VERSION <version>]
-        [FLAGS <flag1> ... <flagN>]
-)
-```
 
 A helper function which uses [cbindgen] to generate C/C++ bindings for a Rust crate.
 If `cbindgen` is not in `PATH` the helper function will automatically try to download
 `cbindgen` and place the built binary into `CMAKE_BINARY_DIR`. The binary is shared
 between multiple invocations of this function.
 
+The function comes with two different signatures. It's recommended to use the `TARGET` based signature when possible.
+
+### Auto mode (With a Rust target imported by corrosion)
+```cmake
+corrosion_experimental_cbindgen(
+        TARGET <imported_target_name>
+        HEADER_NAME <output_header_name>
+        [CBINDGEN_VERSION <version>]
+        [FLAGS <flag1> ... <flagN>]
+)
+```
+
+### Auto-mode specific Arguments
+
 
 * **TARGET**: The name of an imported Rust library target, for which bindings should be generated.
-              If the target was not previously imported by Corrosion, because the crate only produces an
-              `rlib`, you must additionally specify `MANIFEST_DIRECTORY`.
+              If the target is not imported by Corrosion, because the crate only produces an
+              `rlib`, you can instead use the second signature and manually pass `MANIFEST_DIRECTORY`,
+              `CARGO_PACKAGE` and `BINDINGS_TARGET`
 
-* **MANIFEST_DIRECTORY**: Directory of the package defining the library crate bindings should be generated for.
+### Manual mode (Without a Rust target imported by corrosion)
+```cmake
+corrosion_experimental_cbindgen(
+        MANIFEST_DIRECTORY <package_manifest_directory>
+        CARGO_PACKAGE <package_name>
+        BINDINGS_TARGET <cmake_library>
+        [TARGET_TRIPLE <rust_target_triple>]
+        HEADER_NAME <output_header_name>
+        [CBINDGEN_VERSION <version>]
+        [FLAGS <flag1> ... <flagN>]
+)
+```
+
+### Manual-mode specific Arguments
+
+* **MANIFEST_DIRECTORY**: Manual mode only.
+    Directory of the package defining the library crate bindings should be generated for.
     If you want to avoid specifying `MANIFEST_DIRECTORY` you could add a `staticlib` target to your package
     manifest as a workaround to make corrosion import the crate.
+
+* **CARGO_PACKAGE**: Manual mode only.
+    The name of the cargo package that bindings should be generated for.
+    Note: This corresponds to the `cbindgen` `--crate` option, which actually wants a package name.
+
+* **BINDINGS_TARGET**: Manual mode only.
+    Name of an `INTERFACE` CMake target that the generated bindings should be attached to.
+    In auto mode, the generated headers will be attached to the imported rust CMake crate,
+    and corrosion will take care of adding the necessary build dependencies.
+    In manual mode, this target likely doesn't exist, so the user needs to specify an INTERFACE CMake
+    target, which the header files should be attached to. The user must create this target themselves and
+    ensure to add any necessary dependencies (e.g. via `add_dependencies()`) to ensure that consumers of the
+    `INTERFACE` library are not linked before the Rust library has been built.
+
+* **TARGET_TRIPLE**: Manual mode only.
+    Rust target triple (e.g. `x86_64-unknown-linux-gnu`) that cbindgen should use when generating the bindings.
+    Defaults to target triple that corrosion was confiured to compile for.
+
+### Common Arguments
 
 * **HEADER_NAME**: The name of the generated header file. This will be the name which you include in your C/C++ code
                     (e.g. `#include "myproject/myheader.h" if you specify `HEADER_NAME "myproject/myheader.h"`.
 * **CBINDGEN_VERSION**: Version requirement for cbindgen. Exact semantics to be specified. Currently not implemented.
 * **FLAGS**: Arbitrary other flags for `cbindgen`. Run `cbindgen --help` to see the possible flags.
 
-[cbindgen]: https://github.com/eqrion/cbindgen
+[cbindgen]: https://github.com/mozilla/cbindgen
 
 ### Current limitations
 
@@ -1810,12 +2008,16 @@ function(corrosion_experimental_cbindgen)
     set(ONE_VALUE_KEYWORDS
             TARGET
             MANIFEST_DIRECTORY
+            CARGO_PACKAGE
+            BINDINGS_TARGET
+            TARGET_TRIPLE
             HEADER_NAME
-            CBINDGEN_VERSION)
+            CBINDGEN_VERSION
+    )
     set(MULTI_VALUE_KEYWORDS "FLAGS")
     cmake_parse_arguments(PARSE_ARGV 0 CCN "${OPTIONS}" "${ONE_VALUE_KEYWORDS}" "${MULTI_VALUE_KEYWORDS}")
 
-    set(required_keywords TARGET HEADER_NAME)
+    set(required_keywords HEADER_NAME)
     foreach(keyword ${required_keywords})
         if(NOT DEFINED "CCN_${keyword}")
             message(FATAL_ERROR "Missing required parameter `${keyword}`.")
@@ -1823,37 +2025,65 @@ function(corrosion_experimental_cbindgen)
             message(FATAL_ERROR "Required parameter `${keyword}` may not be set to an empty string.")
         endif()
     endforeach()
-    set(rust_target "${CCN_TARGET}")
+    if(NOT (DEFINED CCN_TARGET
+            OR (DEFINED CCN_MANIFEST_DIRECTORY AND DEFINED CCN_BINDINGS_TARGET
+                AND DEFINED CCN_BINDINGS_TARGET)
+            )
+    )
+        message(FATAL_ERROR "Unknown signature for corrosion_experimental_cbindgen.\n"
+                "Either the `TARGET` or the `MANIFEST_DIRECTORY` based signature must be chosen.\n"
+                "Please view the documentation for details on the function signature.\n"
+                "Passed arguments where: `${ARGV}`"
+        )
+    endif()
+
+    if(DEFINED CCN_UNPARSED_ARGUMENTS)
+        message(AUTHOR_WARNING "corrosion_experimental_cbindgen was called with the following unknown arguments: "
+                "`${CCN_UNPARSED_ARGUMENTS}`\n"
+                "Unknown arguments will be ignored."
+        )
+    endif()
     unset(package_manifest_dir)
 
 
-    set(hostbuild_override "$<BOOL:$<TARGET_PROPERTY:${rust_target},${_CORR_PROP_HOST_BUILD}>>")
-    set(cbindgen_target_triple "$<IF:${hostbuild_override},${_CORROSION_RUST_CARGO_HOST_TARGET},${_CORROSION_RUST_CARGO_TARGET}>")
+    if(TARGET "${CCN_TARGET}")
+        set(cbindgen_bindings_target "${CCN_TARGET}")
+        set(hostbuild_override "$<BOOL:$<TARGET_PROPERTY:${CCN_TARGET},${_CORR_PROP_HOST_BUILD}>>")
+        set(cbindgen_target_triple "$<IF:${hostbuild_override},${_CORROSION_RUST_CARGO_HOST_TARGET},${_CORROSION_RUST_CARGO_TARGET}>")
 
-    if(TARGET "${rust_target}")
-        get_target_property(package_manifest_path "${rust_target}" INTERFACE_COR_PACKAGE_MANIFEST_PATH)
+        get_target_property(package_manifest_path "${CCN_TARGET}" INTERFACE_COR_PACKAGE_MANIFEST_PATH)
         if(NOT EXISTS "${package_manifest_path}")
             message(FATAL_ERROR "Internal error: No package manifest found at ${package_manifest_path}")
         endif()
         get_filename_component(package_manifest_dir "${package_manifest_path}" DIRECTORY)
+        get_target_property(rust_cargo_package "${CCN_TARGET}" COR_CARGO_PACKAGE_NAME )
+        if(NOT rust_cargo_package)
+            message(FATAL_ERROR "Internal Error: Could not determine cargo package name for cbindgen. ")
+        endif()
         # todo: as an optimization we could cache the cargo metadata output (but --no-deps makes that slightly more complicated)
     else()
         if(NOT DEFINED CCN_MANIFEST_DIRECTORY)
             message(FATAL_ERROR
-                "`${rust_target}` is not a target imported by corrosion and `MANIFEST_DIRECTORY` was not provided."
-            )
+                "Internal error: There should have been a fatal error already if neither TARGET or "
+                    "MANIFEST_DIRECTORY are specfied.")
+        endif()
+        cmake_path(ABSOLUTE_PATH CCN_MANIFEST_DIRECTORY NORMALIZE OUTPUT_VARIABLE package_manifest_dir)
+        if(DEFINED CCN_TARGET_TRIPLE)
+            set(cbindgen_target_triple "${CCN_TARGET_TRIPLE}")
         else()
-            set(package_manifest_dir "${CCN_MANIFEST_DIRECTORY}")
+            set(cbindgen_target_triple "${Rust_CARGO_TARGET}")
+        endif()
+        set(rust_cargo_package "${CCN_CARGO_PACKAGE}")
+        set(cbindgen_bindings_target "${CCN_BINDINGS_TARGET}")
+        get_target_property(type "${cbindgen_bindings_target}" TYPE)
+        if(NOT ${type} STREQUAL "INTERFACE_LIBRARY")
+            message(AUTHOR_WARNING "The CMake target for the cbindgen generated files is expected to be"
+                " an `INTERFACE` library, but was `${type}` instead."
+            )
         endif()
     endif()
 
-    get_target_property(rust_cargo_package "${rust_target}" COR_CARGO_PACKAGE_NAME )
-    if(NOT rust_cargo_package)
-        message(FATAL_ERROR "Internal Error: Could not determine cargo package name for cbindgen. "
-        )
-    endif()
-    message(STATUS "Using package ${rust_cargo_package} as crate for cbindgen")
-
+    message(STATUS "Using package `${rust_cargo_package}` as crate for cbindgen")
 
     set(output_header_name "${CCN_HEADER_NAME}")
 
@@ -1869,17 +2099,21 @@ function(corrosion_experimental_cbindgen)
             set(executable_postfix ".exe")
         endif()
         set(cbindgen "${local_cbindgen_install_dir}/bin/cbindgen${executable_postfix}")
+
         if(NOT TARGET "_corrosion_cbindgen")
             file(MAKE_DIRECTORY "${local_cbindgen_install_dir}")
+
             add_custom_command(OUTPUT "${cbindgen}"
                 COMMAND ${CMAKE_COMMAND}
                 -E env
-                "CARGO_BUILD_RUSTC=${_CORROSION_RUSTC}"
-                ${_CORROSION_CARGO} install
+                "CARGO_BUILD_RUSTC=$CACHE{CORROSION_TOOLS_RUSTC}"
+                $CACHE{CORROSION_TOOLS_CARGO} install
                     cbindgen
+                    --locked
                     --root "${local_cbindgen_install_dir}"
                     ${_CORROSION_QUIET_OUTPUT_FLAG}
-                COMMENT "Building cbindgen"
+                COMMENT "Building cbindgen with Rust toolchain $CACHE{CORROSION_TOOLS_RUST_TOOLCHAIN}"
+                VERBATIM
                 )
             add_custom_target("_corrosion_cbindgen"
                 DEPENDS "${cbindgen}"
@@ -1888,14 +2122,14 @@ function(corrosion_experimental_cbindgen)
     endif()
 
     set(corrosion_generated_dir "${CMAKE_CURRENT_BINARY_DIR}/corrosion_generated")
-    set(generated_dir "${corrosion_generated_dir}/cbindgen/${rust_target}")
-    set(header_placement_dir "${generated_dir}/include/")
+    set(generated_dir "${corrosion_generated_dir}/cbindgen/${cbindgen_bindings_target}")
+    set(header_placement_dir "${generated_dir}/include")
     set(depfile_placement_dir "${generated_dir}/depfile")
     set(generated_depfile "${depfile_placement_dir}/${output_header_name}.d")
     set(generated_header "${header_placement_dir}/${output_header_name}")
-    message(STATUS "rust target is ${rust_target}")
+
     if(CMAKE_VERSION VERSION_GREATER_EQUAL "3.23")
-        target_sources(${rust_target}
+        target_sources(${cbindgen_bindings_target}
             INTERFACE
             FILE_SET HEADERS
             BASE_DIRS "${header_placement_dir}"
@@ -1903,7 +2137,7 @@ function(corrosion_experimental_cbindgen)
         )
     else()
         # Note: not clear to me how install would best work before CMake 3.23
-        target_include_directories(${rust_target}
+        target_include_directories(${cbindgen_bindings_target}
             INTERFACE
             $<BUILD_INTERFACE:${header_placement_dir}>
             $<INSTALL_INTERFACE:include>
@@ -1916,7 +2150,6 @@ function(corrosion_experimental_cbindgen)
     file(MAKE_DIRECTORY "${generated_header_dir}")
 
     unset(depfile_cbindgen_arg)
-    unset(depfile_cmake_arg)
     get_filename_component(generated_depfile_dir "${generated_depfile}" DIRECTORY)
     file(MAKE_DIRECTORY "${generated_depfile_dir}")
     set(depfile_cbindgen_arg "--depfile=${generated_depfile}")
@@ -1927,6 +2160,9 @@ function(corrosion_experimental_cbindgen)
         COMMAND
         "${CMAKE_COMMAND}" -E env
             TARGET="${cbindgen_target_triple}"
+            # cbindgen invokes cargo-metadata and checks the CARGO environment variable
+            CARGO="${_CORROSION_CARGO}"
+            RUSTC="${_CORROSION_RUSTC}"
             "${cbindgen}"
                     --output "${generated_header}"
                     --crate "${rust_cargo_package}"
@@ -1946,19 +2182,22 @@ function(corrosion_experimental_cbindgen)
         )
     endif()
 
-    if(NOT TARGET "_corrosion_cbindgen_${rust_target}_bindings")
-        add_custom_target(_corrosion_cbindgen_${rust_target}_bindings
+    if(NOT TARGET "_corrosion_cbindgen_${cbindgen_bindings_target}_bindings")
+        add_custom_target(_corrosion_cbindgen_${cbindgen_bindings_target}_bindings
                 COMMENT "Generate cbindgen bindings for package ${rust_cargo_package}"
         )
     endif()
     # Users might want to call cbindgen multiple times, e.g. to generate separate C++ and C header files.
     string(MAKE_C_IDENTIFIER "${output_header_name}" header_identifier )
-    add_custom_target("_corrosion_cbindgen_${rust_target}_bindings_${header_identifier}"
+    add_custom_target("_corrosion_cbindgen_${cbindgen_bindings_target}_bindings_${header_identifier}"
             DEPENDS "${generated_header}"
-            COMMENT "Generate ${generated_header} for ${rust_target}"
+            COMMENT "Generate ${generated_header} for ${cbindgen_bindings_target}"
     )
-    add_dependencies("_corrosion_cbindgen_${rust_target}_bindings" "_corrosion_cbindgen_${rust_target}_bindings_${header_identifier}")
-    add_dependencies(${rust_target} "_corrosion_cbindgen_${rust_target}_bindings")
+    add_dependencies("_corrosion_cbindgen_${cbindgen_bindings_target}_bindings" "_corrosion_cbindgen_${cbindgen_bindings_target}_bindings_${header_identifier}")
+    add_dependencies(${cbindgen_bindings_target} "_corrosion_cbindgen_${cbindgen_bindings_target}_bindings")
+    if(TARGET "${CCN_TARGET}")
+        add_dependencies(cargo-build_${CCN_TARGET} "_corrosion_cbindgen_${cbindgen_bindings_target}_bindings")
+    endif()
 endfunction()
 
 # Parse the version of a Rust package from it's package manifest (Cargo.toml)