+++ tools/corrosion/cmake/FindRust.cmake
@@ -334,18 +334,18 @@ if (Rust_RESOLVE_RUSTUP_TOOLCHAINS)
     set(_DISCOVERED_TOOLCHAINS_VERSION "")
 
     foreach(_TOOLCHAIN_RAW ${_TOOLCHAINS_RAW})
-        if (_TOOLCHAIN_RAW MATCHES "([a-zA-Z0-9\\._\\-]+)[ \t\r\n]?(\\(active, default\\)|\\(default\\) \\(override\\)|\\(default\\)|\\(override\\))?[ \t\r\n]+(.+)")
+        if (_TOOLCHAIN_RAW MATCHES "([a-zA-Z0-9\\._\\-]+)[ \t\r\n]?(\\(active\\)|\\(active, default\\)|\\(default\\) \\(override\\)|\\(default\\)|\\(override\\))?[ \t\r\n]+(.+)")
             set(_TOOLCHAIN "${CMAKE_MATCH_1}")
             set(_TOOLCHAIN_TYPE "${CMAKE_MATCH_2}")
 
             set(_TOOLCHAIN_PATH "${CMAKE_MATCH_3}")
             set(_TOOLCHAIN_${_TOOLCHAIN}_PATH "${CMAKE_MATCH_3}")
 
-            if (_TOOLCHAIN_TYPE MATCHES ".*\\(default\\).*")
+            if (_TOOLCHAIN_TYPE MATCHES ".*\\((active, )?default\\).*")
                 set(_TOOLCHAIN_DEFAULT "${_TOOLCHAIN}")
             endif()
 
-            if (_TOOLCHAIN_TYPE MATCHES ".*\\(override\\).*")
+            if (_TOOLCHAIN_TYPE MATCHES ".*\\((active|override)\\).*")
                 set(_TOOLCHAIN_OVERRIDE "${_TOOLCHAIN}")
             endif()
 
@@ -368,7 +368,11 @@ if (Rust_RESOLVE_RUSTUP_TOOLCHAINS)
                 else()
                     set(_TOOLCHAIN_${_TOOLCHAIN}_IS_NIGHTLY "FALSE")
                 endif()
-                if(EXISTS "${_TOOLCHAIN_PATH}/bin/cargo")
+                set(_suffix "")
+                if(CMAKE_HOST_WIN32)
+                    set(_suffix ".exe")
+                endif()
+                if(EXISTS "${_TOOLCHAIN_PATH}/bin/cargo${_suffix}")
                     list(APPEND _DISCOVERED_TOOLCHAINS_CARGO_PATH "${_TOOLCHAIN_PATH}/bin/cargo")
                 else()
                     list(APPEND _DISCOVERED_TOOLCHAINS_CARGO_PATH "NOTFOUND")
@@ -376,7 +380,7 @@ if (Rust_RESOLVE_RUSTUP_TOOLCHAINS)
             else()
                 message(AUTHOR_WARNING "Unexpected output from `rustc --version` for Toolchain `${_TOOLCHAIN}`: "
                         "`${_TOOLCHAIN_RAW_VERSION}`.\n"
-                        "Ignoring this toolchain."
+                        "Ignoring this toolchain (Path: `${_TOOLCHAIN_PATH}`)."
                 )
             endif()
         else()